requirements.txt
google-generativeai
streamlit
Pillow
python-dotenv